fix(@findify/react-components): move display logic inside child components for easier access
