+++ packages/react-components/src/layouts/Autocomplete/Fullscreen/index.tsx
@@ -8,11 +8,21 @@ import cx from 'classnames';
 import { useSuggestions } from '@findify/react-connect';
 import { Immutable } from '@findify/store-configuration';
 import useTranslations from 'helpers/useTranslations';
+import { AutocompleteType } from '../types';
 
 /** TODO: NEED TO REFACTOR  */
-export default ({ theme = styles, showSuggestions }) => {
-  const { suggestions } = useSuggestions<Immutable.AutocompleteConfig>();
+export default ({ theme = styles, isMobile }) => {
+  const { suggestions, config } = useSuggestions<Immutable.AutocompleteConfig>();
   const translate = useTranslations();
+
+  const viewType: AutocompleteType = isMobile
+    ? config.getIn(['template', 'mobile'])
+    : config.getIn(['template', 'desktop']);
+
+  const templateSetting = config.get(viewType);
+
+  const showSuggestions = !isMobile || !!templateSetting?.getIn(['suggestions', 'display']);
+
   return (
     <div
       display-if={showSuggestions &&  suggestions && suggestions.size > 0}
